Use #Preview macros and split bank account tests into their own files
Addresses review feedback: convert AddEditBankAccountItemView previews to #Preview macros, and move bank account processor and view inspector tests into dedicated files.
