+++ BitwardenShared/UI/Vault/VaultItem/AddEditItem/AddEditItemProcessorTests.swift
@@ -2193,120 +2193,6 @@ class AddEditItemProcessorTests: BitwardenTestCase {
         XCTAssertFalse(subject.state.driversLicenseItemState.isLicenseNumberVisible)
     }
 
-    /// `receive(_:)` with `.bankAccountFieldChanged(.bankNameChanged)` updates the state correctly.
-    @MainActor
-    func test_receive_bankAccountFieldChanged_bankNameChanged() {
-        subject.state.bankAccountItemState.bankName = "Old Bank"
-        subject.receive(.bankAccountFieldChanged(.bankNameChanged("Bank of America")))
-        XCTAssertEqual(subject.state.bankAccountItemState.bankName, "Bank of America")
-    }
-
-    /// `receive(_:)` with `.bankAccountFieldChanged(.nameOnAccountChanged)` updates the state correctly.
-    @MainActor
-    func test_receive_bankAccountFieldChanged_nameOnAccountChanged() {
-        subject.state.bankAccountItemState.nameOnAccount = "Old Name"
-        subject.receive(.bankAccountFieldChanged(.nameOnAccountChanged("Personal Checking")))
-        XCTAssertEqual(subject.state.bankAccountItemState.nameOnAccount, "Personal Checking")
-    }
-
-    /// `receive(_:)` with `.bankAccountFieldChanged(.accountTypeChanged)` updates the state correctly.
-    @MainActor
-    func test_receive_bankAccountFieldChanged_accountTypeChanged() {
-        subject.state.bankAccountItemState.accountType = .default
-        subject.receive(.bankAccountFieldChanged(.accountTypeChanged(.custom(.checking))))
-        XCTAssertEqual(subject.state.bankAccountItemState.accountType, .custom(.checking))
-    }
-
-    /// `receive(_:)` with `.bankAccountFieldChanged(.accountNumberChanged)` updates the state correctly.
-    @MainActor
-    func test_receive_bankAccountFieldChanged_accountNumberChanged() {
-        subject.state.bankAccountItemState.accountNumber = "111"
-        subject.receive(.bankAccountFieldChanged(.accountNumberChanged("1234567890123456")))
-        XCTAssertEqual(subject.state.bankAccountItemState.accountNumber, "1234567890123456")
-    }
-
-    /// `receive(_:)` with `.bankAccountFieldChanged(.routingNumberChanged)` updates the state correctly.
-    @MainActor
-    func test_receive_bankAccountFieldChanged_routingNumberChanged() {
-        subject.state.bankAccountItemState.routingNumber = "111"
-        subject.receive(.bankAccountFieldChanged(.routingNumberChanged("1234567890")))
-        XCTAssertEqual(subject.state.bankAccountItemState.routingNumber, "1234567890")
-    }
-
-    /// `receive(_:)` with `.bankAccountFieldChanged(.branchNumberChanged)` updates the state correctly.
-    @MainActor
-    func test_receive_bankAccountFieldChanged_branchNumberChanged() {
-        subject.state.bankAccountItemState.branchNumber = "1"
-        subject.receive(.bankAccountFieldChanged(.branchNumberChanged("100")))
-        XCTAssertEqual(subject.state.bankAccountItemState.branchNumber, "100")
-    }
-
-    /// `receive(_:)` with `.bankAccountFieldChanged(.pinChanged)` updates the state correctly.
-    @MainActor
-    func test_receive_bankAccountFieldChanged_pinChanged() {
-        subject.state.bankAccountItemState.pin = "0000"
-        subject.receive(.bankAccountFieldChanged(.pinChanged("1234")))
-        XCTAssertEqual(subject.state.bankAccountItemState.pin, "1234")
-    }
-
-    /// `receive(_:)` with `.bankAccountFieldChanged(.swiftCodeChanged)` updates the state correctly.
-    @MainActor
-    func test_receive_bankAccountFieldChanged_swiftCodeChanged() {
-        subject.state.bankAccountItemState.swiftCode = "000"
-        subject.receive(.bankAccountFieldChanged(.swiftCodeChanged("123234")))
-        XCTAssertEqual(subject.state.bankAccountItemState.swiftCode, "123234")
-    }
-
-    /// `receive(_:)` with `.bankAccountFieldChanged(.ibanChanged)` updates the state correctly.
-    @MainActor
-    func test_receive_bankAccountFieldChanged_ibanChanged() {
-        subject.state.bankAccountItemState.iban = "000"
-        subject.receive(.bankAccountFieldChanged(.ibanChanged("23423434543")))
-        XCTAssertEqual(subject.state.bankAccountItemState.iban, "23423434543")
-    }
-
-    /// `receive(_:)` with `.bankAccountFieldChanged(.bankContactPhoneChanged)` updates the state correctly.
-    @MainActor
-    func test_receive_bankAccountFieldChanged_bankContactPhoneChanged() {
-        subject.state.bankAccountItemState.bankContactPhone = "000"
-        subject.receive(.bankAccountFieldChanged(.bankContactPhoneChanged("123-456-7890")))
-        XCTAssertEqual(subject.state.bankAccountItemState.bankContactPhone, "123-456-7890")
-    }
-
-    /// `receive(_:)` with `.bankAccountFieldChanged(.toggleAccountNumberVisibilityChanged)` updates
-    /// the state correctly.
-    @MainActor
-    func test_receive_bankAccountFieldChanged_toggleAccountNumberVisibilityChanged() {
-        subject.state.bankAccountItemState.isAccountNumberVisible = false
-        subject.receive(.bankAccountFieldChanged(.toggleAccountNumberVisibilityChanged(true)))
-        XCTAssertTrue(subject.state.bankAccountItemState.isAccountNumberVisible)
-
-        subject.receive(.bankAccountFieldChanged(.toggleAccountNumberVisibilityChanged(false)))
-        XCTAssertFalse(subject.state.bankAccountItemState.isAccountNumberVisible)
-    }
-
-    /// `receive(_:)` with `.bankAccountFieldChanged(.togglePinVisibilityChanged)` updates the state correctly.
-    @MainActor
-    func test_receive_bankAccountFieldChanged_togglePinVisibilityChanged() {
-        subject.state.bankAccountItemState.isPinVisible = false
-        subject.receive(.bankAccountFieldChanged(.togglePinVisibilityChanged(true)))
-        XCTAssertTrue(subject.state.bankAccountItemState.isPinVisible)
-
-        subject.receive(.bankAccountFieldChanged(.togglePinVisibilityChanged(false)))
-        XCTAssertFalse(subject.state.bankAccountItemState.isPinVisible)
-    }
-
-    /// `receive(_:)` with `.bankAccountFieldChanged(.toggleIbanVisibilityChanged)` updates the state correctly.
-    @MainActor
-    func test_receive_bankAccountFieldChanged_toggleIbanVisibilityChanged() {
-        subject.state.bankAccountItemState.isIbanVisible = false
-        subject.receive(.bankAccountFieldChanged(.toggleIbanVisibilityChanged(true)))
-        XCTAssertTrue(subject.state.bankAccountItemState.isIbanVisible)
-
-        subject.receive(.bankAccountFieldChanged(.toggleIbanVisibilityChanged(false)))
-        XCTAssertFalse(subject.state.bankAccountItemState.isIbanVisible)
-    }
-
     /// `receive(_:)` with `.identityFieldChanged(.titleChanged)` with a value updates the state correctly.
     @MainActor
     func test_receive_identity_titleChange_withValidValue() {
